refactor: Formula.java 유효성 검사 로직 추가 1. 계산식 도메인에 책임 추가 2. 별도로 만든 유효성 검사 제거
